+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Removed
 - Dropped Python 3.10 support
 - Unused and obsolete serial_images_nexus for copying tristan nexus files as multi_image is used.
+- Standalone serial Tristan writer as now integrated with I19-2 writers.
 
 ### Changed
 - Tidier parameter model for I19-2.